pyproject.toml
[project]
name = "ytclfr"
version = "0.1.0"
description = "YouTube to structured knowledge pipeline using OCR and AI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "celery>=5.4.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "paddleocr>=2.8.1",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "prometheus-client>=0.21.0",
  "python-dotenv>=1.0.1",
  "rapidfuzz>=3.9.7",
  "sentry-sdk>=2.14.0",
  "psycopg[binary]>=3.2.1",
  "redis>=5.0.8",
  "scenedetect>=0.6.4",
  "sqlalchemy>=2.0.32",
  "uvicorn[standard]>=0.30.6",
  "yt-dlp>=2025.1.26"
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
